requirements.txt
opencv-python-headless
numpy
Pillow
fastapi
uvicorn
python-multipart
jinja2
python-dotenv
